PR4 corrections
