--- src/main/java/com/github/boscojared/disparse/discord/Helpable.java
@@ -1,14 +1,0 @@
-package com.github.boscojared.disparse.discord;
-
-import com.github.boscojared.disparse.parser.Command;
-import com.github.boscojared.disparse.parser.CommandFlag;
-import java.util.Collection;
-
-public interface Helpable<E> {
-    void help(E event, Command command, Collection<CommandFlag> flags, Collection<Command> commands);
-    void allCommands(E event, Collection<Command> commands);
-    void setPrefix(String prefix);
-    void commandNotFound(E event, String userInput);
-    void roleNotMet(E event, Command command);
-    void optionRequired(E event, String message);
-}
