--- src/main/java/com/github/boscojared/disparse/parser/Types.java
@@ -1,5 +1,0 @@
-package com.github.boscojared.disparse.parser;
-
-public enum Types {
-    INT, STR, BOOL, LIST
-}
