--- src/main/java/com/github/boscojared/disparse/parser/exceptions/OptionRequired.java
@@ -1,5 +1,0 @@
-package com.github.boscojared.disparse.parser.exceptions;
-
-public class OptionRequired extends RuntimeException {
-    public OptionRequired(String message) { super(message); }
-}
