--- src/main/java/com/github/boscojared/disparse/parser/reflection/CommandHandler.java
@@ -1,15 +1,0 @@
-package com.github.boscojared.disparse.parser.reflection;
-
-import java.lang.annotation.ElementType;
-import java.lang.annotation.Retention;
-import java.lang.annotation.RetentionPolicy;
-import java.lang.annotation.Target;
-
-@Retention(RetentionPolicy.RUNTIME)
-@Target(ElementType.METHOD)
-public @interface CommandHandler {
-    String commandName();
-    String description() default "no description available";
-
-    String[] roles() default {};
-}
