--- src/main/java/com/github/boscojared/disparse/utils/Shlex.java
@@ -1,39 +1,0 @@
-package com.github.boscojared.disparse.utils;
-
-import java.util.ArrayList;
-import java.util.List;
-
-public class Shlex {
-    public static List<String> shlex(CharSequence argString) {
-        List<String> tokens = new ArrayList<>();
-        boolean escaping = false;
-        char quoteChar = ' ';
-        boolean quoting = false;
-        StringBuilder current = new StringBuilder();
-        for (int i = 0; i < argString.length(); i++) {
-            char c = argString.charAt(i);
-            if (escaping) {
-                current.append(c);
-                escaping = false;
-            } else if (c == '\\' && !(quoting && quoteChar == '\'')) {
-                escaping = true;
-            } else if (quoting && c == quoteChar) {
-                quoting = false;
-            } else if (!quoting && (c == '\'' || c == '"')) {
-                quoting = true;
-                quoteChar = c;
-            } else if (!quoting && Character.isWhitespace(c)) {
-                if (current.length() > 0) {
-                    tokens.add(current.toString());
-                    current = new StringBuilder();
-                }
-            } else {
-                current.append(c);
-            }
-        }
-        if (current.length() > 0) {
-            tokens.add(current.toString());
-        }
-        return tokens;
-    }
-}
