+++ src/main/java/disparse/parser/Types.java
@@ -1,0 +1,5 @@
+package disparse.parser;
+
+public enum Types {
+  INT, STR, BOOL, LIST
+}
