+++ src/main/java/disparse/parser/exceptions/NoCommandNameFound.java
@@ -1,0 +1,8 @@
+package disparse.parser.exceptions;
+
+public class NoCommandNameFound extends RuntimeException {
+
+  public NoCommandNameFound(String message) {
+    super(message);
+  }
+}
