+++ src/main/java/disparse/parser/exceptions/OptionRequired.java
@@ -1,0 +1,6 @@
+package disparse.parser.exceptions;
+
+public class OptionRequired extends RuntimeException {
+
+  public OptionRequired(String message) { super(message); }
+}
