+++ src/main/java/disparse/parser/exceptions/OptionRequiresValue.java
@@ -1,0 +1,8 @@
+package disparse.parser.exceptions;
+
+public class OptionRequiresValue extends RuntimeException {
+
+  public OptionRequiresValue(String message) {
+    super(message);
+  }
+}
