+++ src/main/java/disparse/parser/reflection/ParsedEntity.java
@@ -1,0 +1,11 @@
+package disparse.parser.reflection;
+
+import java.lang.annotation.ElementType;
+import java.lang.annotation.Retention;
+import java.lang.annotation.RetentionPolicy;
+import java.lang.annotation.Target;
+
+@Retention(RetentionPolicy.RUNTIME)
+@Target(ElementType.TYPE)
+public @interface ParsedEntity {
+}
